+++ tasks/html.js
@@ -1,0 +1,7 @@
+var copy = require('./util').copy
+
+function html(callback) {
+  copy('./app/index.html', './build/', callback)
+}
+
+module.exports = html
